# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/user/create": {
            "post": {
                "description": "用户注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "用户注册",
                "operationId": "CreateUser",
                "parameters": [
                    {
                        "description": "新用户的详细信息",
                        "name": "RequestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.CreateUserRole"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "成功创建的用户信息",
                        "schema": {
                            "$ref": "#/definitions/common.CommonResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数有误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/v1/user/login": {
            "post": {
                "description": "用户登陆",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户登陆"
                ],
                "summary": "用户登陆",
                "operationId": "UserLogin",
                "parameters": [
                    {
                        "description": "登陆用户信息信息",
                        "name": "RequestBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UserLogin"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "用户信息",
                        "schema": {
                            "$ref": "#/definitions/user.UserLoginResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数有误",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.CommonResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "user.CreateUserRole": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "role_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "user.UserInfo": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "user.UserLogin": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "user.UserLoginResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/user.UserInfo"
                }
            }
        }
    }
}